Rename novantExt -> novantMercuryExt; drop auto-sync
Rework this extension to work side-by-side with the new Apollo-only novantExt.
This required renaming all tags/ext/funcs to use `novantMercury` prefix in
order to distinguish.

This changeset also drops the problematic "auto-his-sync" behavoir in favor
of traditional per-point sync.  This is much less effecient use of the API
but is much more robust with how SkySpark is designed and more consistent
with local workflows and conventions.